requirements.txt
fastapi[standard]
uvicorn>=0.29.0,<0.30.0
python-dotenv>=1.0.1,<2.0.0
requests>=2.31.0,<3.0.0
motor
databases[postgresql]
asyncpg